修复了部分车险不能获取关联车险的问题

+++ src/main/java/org/example/crm/entity/car/CarInsurance.java
@@ -15,7 +15,7 @@ public class CarInsurance {
     /**
      * 车险id
      */
-    private Integer productId;
+    private Long productId;
 
     /**
      * 保险名称
@@ -61,5 +61,6 @@ public class CarInsurance {
      * 保险的好处
      */
     private String benefits;
+
 }
